Add files via upload
Created during my junior and senior years of high school (I'm a senior right now) and during that period I enrolled in the AP Computer Science A course. Includes one Python file as well, because we experimented with the language as well during junior year, in preparation for the easyCTF Hackathon.